// src/generated/models/Jsi_stagesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Jsi_stagesjsi_scope = {
  100000000: 'General',
  100000001: 'AreaSpecific'
} as const;
export type Jsi_stagesjsi_scope = keyof typeof Jsi_stagesjsi_scope;
export const Jsi_stagesstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Jsi_stagesstatecode = keyof typeof Jsi_stagesstatecode;
export const Jsi_stagesstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Jsi_stagesstatuscode = keyof typeof Jsi_stagesstatuscode;

export interface Jsi_stagesBase {
  importsequencenumber?: number;
  jsi_businessdaysdeadline?: number;
  jsi_name?: string;
  jsi_order: number;
  jsi_scope: Jsi_stagesjsi_scope;
  jsi_stageid: string;
  overriddencreatedon?: string;
  statecode: Jsi_stagesstatecode;
  statuscode?: Jsi_stagesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Jsi_stages extends Jsi_stagesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  jsi_scopename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}