// src/generated/models/Jsi_documentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Jsi_documentsjsi_criticality = {
  100000000: 'High',
  100000001: 'Medium',
  100000002: 'Low'
} as const;
export type Jsi_documentsjsi_criticality = keyof typeof Jsi_documentsjsi_criticality;
export const Jsi_documentsjsi_status = {
  100000000: 'Current',
  100000001: 'Updated',
  100000002: 'Draft'
} as const;
export type Jsi_documentsjsi_status = keyof typeof Jsi_documentsjsi_status;
export const Jsi_documentsstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Jsi_documentsstatecode = keyof typeof Jsi_documentsstatecode;
export const Jsi_documentsstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Jsi_documentsstatuscode = keyof typeof Jsi_documentsstatuscode;

export interface Jsi_documentsBase {
  importsequencenumber?: number;
  "jsi_Category@odata.bind"?: string;
  jsi_code: string;
  jsi_criticality: Jsi_documentsjsi_criticality;
  jsi_documentid: string;
  jsi_mandatory?: boolean;
  jsi_owningdepartment?: string;
  jsi_pagecount?: number;
  jsi_readingminutes?: number;
  "jsi_Stage@odata.bind": string;
  jsi_status: Jsi_documentsjsi_status;
  jsi_summary?: string;
  jsi_title?: string;
  jsi_updatedon?: string;
  jsi_version: string;
  overriddencreatedon?: string;
  statecode: Jsi_documentsstatecode;
  statuscode?: Jsi_documentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Jsi_documents extends Jsi_documentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  jsi_categoryname?: string;
  jsi_criticalityname?: string;
  jsi_file?: string;
  jsi_file_name?: string;
  jsi_mandatoryname?: string;
  jsi_stagename?: string;
  jsi_statusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  jsi_category?: object;
  _jsi_category_value?: string;
  jsi_stage?: object;
  _jsi_stage_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}
export type Jsi_documentsUploadColumnName = 'jsi_file';